+++ packages/desktop/src/components/date/index.ts
@@ -1,0 +1,2 @@
+export { default as Calendar } from "./Calendar.svelte";
+export { default as DateField } from "./DateField.svelte";
